astar/requirements.txt
osmnx==2.0.2
networkx>=3.0
numpy>=1.24
shapely>=2.0
psycopg[binary]>=3.1
python-dotenv>=1.0
geopandas>=0.14
geopy>=2.3
scikit-learn>=1.2,<1.5